// apps/web/src/lib/supabase/types.ts
export type Enum_user_role = 'dono' | 'gerente' | 'atendente' | 'cliente';
// ...
export type users_profiles = {
  id: string;
  user_id: string;
  restaurant_id: string;
  role: Enum_user_role;
  name: string;
  email: string;
  created_at: string;
};